Rework image data extraction
The old license plate extraction api no longer exists and therefore a
new one needed to be used, this required some additional changes
as this api only supports files up to 3MB and therefore the images
needed to be compressed beforehand.

At the same time the previous imageData function was split in to three
(extractPlate, extractExifData and createThumbnail) to make the code
more manageable.